refactor: migrate ElectronicBasis to problems module
