+++ qiskit_nature/second_q/transformers/basis_transformer.py
@@ -21,14 +21,13 @@ from typing import cast, TypeVar
 from qiskit_nature.exceptions import QiskitNatureError
 from qiskit_nature.second_q.hamiltonians import ElectronicEnergy, Hamiltonian
 from qiskit_nature.second_q.operators import ElectronicIntegrals, PolynomialTensor
-from qiskit_nature.second_q.problems import BaseProblem, ElectronicStructureProblem
+from qiskit_nature.second_q.problems import BaseProblem, ElectronicBasis, ElectronicStructureProblem
 from qiskit_nature.second_q.properties import (
     AngularMomentum,
     ElectronicDipoleMoment,
     Magnetization,
     ParticleNumber,
 )
-from qiskit_nature.second_q.properties.bases import ElectronicBasis
 
 from .base_transformer import BaseTransformer
 
@@ -56,7 +55,7 @@ class BasisTransformer(BaseTransformer):
     """A transformer to map from one basis to another.
 
     Since problems have a basis associated with them (e.g. the
-    :class:`qiskit_nature.second_q.properties.bases.ElectronicBasis` in the case of the
+    :class:`qiskit_nature.second_q.problems.ElectronicBasis` in the case of the
     :class:`qiskit_nature.second_q.problems.ElectronicStructureProblem`), this transformer can be
     used to map from one :attr:`initial_basis` to another :attr:`final_basis`.
 
@@ -69,7 +68,7 @@ class BasisTransformer(BaseTransformer):
         ao2mo_coeff, ao2mo_coeff_b = ...
 
         from qiskit_nature.second_q.operators import ElectronicIntegrals
-        from qiskit_nature.second_q.properties.bases import ElectronicBasis
+        from qiskit_nature.second_q.problems import ElectronicBasis
         from qiskit_nature.second_q.transformers import BasisTransformer
 
         transformer = BasisTransformer(
